refactor core tests
