+++ projects/core/src/auth/http-interceptors/auth-error.interceptor.spec.ts
@@ -1,10 +1,10 @@
 import {
+  HTTP_INTERCEPTORS,
   HttpClient,
   HttpHandler,
   HttpHeaders,
   HttpParams,
   HttpRequest,
-  HTTP_INTERCEPTORS,
 } from '@angular/common/http';
 import {
   HttpClientTestingModule,
@@ -15,13 +15,9 @@ import { Type } from '@angular/core';
 import { inject, TestBed } from '@angular/core/testing';
 import { Observable, of, throwError } from 'rxjs';
 import { catchError } from 'rxjs/operators';
-import {
-  USE_CLIENT_TOKEN,
-  USE_CUSTOMER_SUPPORT_AGENT_TOKEN,
-} from '../../occ/utils/interceptor-util';
+import { USE_CLIENT_TOKEN } from '../../occ/utils/interceptor-util';
 import { AuthService } from '../facade/auth.service';
 import { ClientErrorHandlingService } from '../services/client-error/client-error-handling.service';
-import { CustomerSupportAgentErrorHandlingService } from '../services/csagent-error/csagent-error-handling.service';
 import { UserErrorHandlingService } from '../services/user-error/user-error-handling.service';
 import { AuthErrorInterceptor } from './auth-error.interceptor';
 
@@ -48,16 +44,11 @@ class MockAuthService {
   logout(): void {}
 }
 
-class MockCustomerSupportAgentErrorHandlingService {
-  terminateCustomerSupportAgentExpiredSession(): void {}
-}
-
 describe('AuthErrorInterceptor', () => {
   let userErrorHandlingService: UserErrorHandlingService;
   let clientErrorHandlingService: ClientErrorHandlingService;
   let authService: AuthService;
   let httpMock: HttpTestingController;
-  let csagentErrorHandlingService: CustomerSupportAgentErrorHandlingService;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
@@ -74,10 +65,6 @@ describe('AuthErrorInterceptor', () => {
         {
           provide: AuthService,
           useClass: MockAuthService,
-        },
-        {
-          provide: CustomerSupportAgentErrorHandlingService,
-          useClass: MockCustomerSupportAgentErrorHandlingService,
         },
         {
           provide: HTTP_INTERCEPTORS,
@@ -97,9 +84,6 @@ describe('AuthErrorInterceptor', () => {
     httpMock = TestBed.get(HttpTestingController as Type<
       HttpTestingController
     >);
-    csagentErrorHandlingService = TestBed.get(
-      CustomerSupportAgentErrorHandlingService
-    );
     spyOn(userErrorHandlingService, 'handleExpiredUserToken').and.returnValue(
       of({} as any)
     );
@@ -108,10 +92,6 @@ describe('AuthErrorInterceptor', () => {
       clientErrorHandlingService,
       'handleExpiredClientToken'
     ).and.returnValue(of({}));
-    spyOn(
-      csagentErrorHandlingService,
-      'terminateCustomerSupportAgentExpiredSession'
-    ).and.stub();
   });
 
   it(`should catch 401 error for a client token`, inject(
@@ -227,38 +207,4 @@ describe('AuthErrorInterceptor', () => {
       { status: 400, statusText: 'Error' }
     );
   }));
-
-  it(`should catch 401 error for a csagent request`, inject(
-    [HttpClient],
-    (http: HttpClient) => {
-      const headers = new HttpHeaders().set(
-        USE_CUSTOMER_SUPPORT_AGENT_TOKEN,
-        'true'
-      );
-      const options = {
-        headers,
-      };
-      http.get('/test', options).subscribe(result => {
-        expect(result).toBeTruthy();
-      });
-
-      const mockReq: TestRequest = httpMock.expectOne(req => {
-        return req.method === 'GET';
-      });
-      mockReq.flush(
-        {
-          errors: [
-            {
-              type: 'InvalidTokenError',
-              message: 'Invalid access token: some token',
-            },
-          ],
-        },
-        { status: 401, statusText: 'Error' }
-      );
-      expect(
-        csagentErrorHandlingService.terminateCustomerSupportAgentExpiredSession
-      ).toHaveBeenCalled();
-    }
-  ));
 });
